back_end/requirements.txt
Django==4.2.7
django_cors_headers==4.3.1
djangorestframework==3.14.0
protobuf==3.19.6
google-cloud-bigquery
psycopg2~=2.9.9